refactor: newsDetail
사용되지 않는 파일 제거

폴더를 만들어서 세분화
