+++ src/components/NewsDetail/Detail.tsx
@@ -8,13 +8,12 @@ import NewsDetailMeta from './NewsDetailMeta';
 import NewsDetailChipList from './Chip/NewsDetailChipList';
 import NoDescription from './NoDescription';
 import NewsDetailContentWrapper from './Content/NewsDetailContentWrapper';
-import NewsDetailContent from './NewsDetailContent';
+import NewsDetailContent from './Content/NewsDetailContent';
 import Button from '../Button/Button';
 import { MdOutlineLink } from 'react-icons/md';
 import { HorizontalLine } from '../UI/DivideLine';
 import OtherNews from './OtherNews';
-import AddCategoryForm from './AddCategoryForm';
-import AddSymbolForm from './AddSymbolForm';
+import { AddCategoryForm, AddSymbolForm } from './Form';
 import TopScrollButton from './TopScrollButton';
 import BenefitBar from './BenefitBar';
 import { type NewsDetailPayload, type NewsDetailSymbol, type UserPayload } from '@/interfaces/Dto/News';
