+++ src/components/NewsDetail/DetailThumbnail.tsx
@@ -4,10 +4,6 @@ interface DetailThumbnailProps {
   src?: string;
 }
 
-export function DetailThumbnailLoading() {
-  return <NewsThumbnailLoading className={'mx-auto aspect-[16/10] !h-auto !w-full max-w-2xl !rounded-lg'} />;
-}
-
 function DetailThumbnail({ src }: DetailThumbnailProps) {
   return (
     <NewsThumbnail
