+++ src/components/NewsDetail/Form/index.ts
@@ -1,0 +1,4 @@
+import AddCategoryForm from './AddCategoryForm';
+import AddSymbolForm from './AddSymbolForm';
+
+export { AddCategoryForm, AddSymbolForm };
