+++ src/components/NewsDetail/Loading/NewsDetailThumbnailLoading.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import { NewsThumbnailLoading } from '@/components/Image/NewsThumbnail';
+
+function NewsDetailThumbnailLoading() {
+  return <NewsThumbnailLoading className={'mx-auto aspect-[16/10] !h-auto !w-full max-w-2xl !rounded-lg'} />;
+}
+
+export default NewsDetailThumbnailLoading;
