+++ src/components/NewsDetail/Loading/NewsDetailTitleLoading.tsx
@@ -1,0 +1,9 @@
+function NewsDetailTitleLoading() {
+  return (
+    <h2 className="skeleton_loading">
+      <p className="h-7 w-80 pc:w-[40rem]" />
+    </h2>
+  );
+}
+
+export default NewsDetailTitleLoading;
