+++ src/components/NewsDetail/NewsDetailTitle.tsx
@@ -8,14 +8,6 @@ interface NewsDetailTitleProps {
   isTranslation: boolean;
 }
 
-export function NewsDetailTitleLoading() {
-  return (
-    <h2 className="skeleton_loading">
-      <p className="h-7 w-80 pc:w-[40rem]" />
-    </h2>
-  );
-}
-
 function NewsDetailTitle({ title, isTranslation }: NewsDetailTitleProps) {
   return <h2 className="break-all text-xl font-semibold">{title[isTranslation ? 'kor' : 'eng']}</h2>;
 }
